[TOKN-61] Break up code into modules.
