+++ style_dictionary/custom/helperFunctions/helperFunctions.js
@@ -1,0 +1,25 @@
+const { cssNoUnits, units } = require('../constants/constants')
+
+// Transform property name from camel case to hyphenated string usable by CSS.
+const formatName = (name) => {
+  const isUpperCase = (string) => /^[A-Z]*$/.test(string)
+  const letters = name.split('')
+  return letters
+    .map((letter) => {
+      if (isUpperCase(letter)) {
+        return `-${letter.toLowerCase()}`
+      } else return letter
+    })
+    .join('')
+}
+
+// Append units to CSS property value if required.
+const formatValue = (name, value) => {
+  if (value === 0) return value
+  if (typeof value !== 'number') return value
+  if (cssNoUnits.includes(formatName(name))) return value
+
+  return value + units
+}
+
+module.exports = { formatName, formatValue }
